Cleaned up project:
Three20 re-added via the recommended ttmodule.py script
Also re-added RestKit after(!) the Three20 script!
Note: As the Header Search Paths for the target is modified by the script.
I manually added here the RestKit path
(if we go by the install instructions we would overwrite the path for Three20)

Signed-off-by: David <d.steinberger@gmx.at>